% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLH.R
\name{GLH}
\alias{GLH}
\title{GLH}
\arguments{
\item{x}{a list of columns of covariates}

\item{y}{continuous outcome}

\item{intercept}{whether to include intercept in the model, the default is True}

\item{c}{contrast matrix,}

\item{alpha}{significance level, the default value is 0.05}

\item{rhs}{constant matrix, the default value  is a scalar zero.}
}
\value{
F statistic, P value
}
\description{
General Linear Hypothesis
}
\examples{
square(3)

}
