% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLH.R
\name{GLH}
\alias{GLH}
\title{GLH}
\usage{
GLH(x, y, intr = TRUE, predict = NULL, contrast = NULL, rhs = 0, alpha = 0.05)
}
\arguments{
\item{x}{a list of columns of covariates}

\item{y}{continuous outcome}

\item{predict}{predict matrix used to estimate the outcomes, the columns numbers should be equal to number of covariates}

\item{contrast}{contrast matrix,}

\item{rhs}{constant matrix, the default value  is a scalar zero.}

\item{alpha}{significance level, the default value is 0.05}

\item{intercept}{whether to include intercept in the model, the default is True}
}
\value{
The coefficients of covariates, and some basic statistics. The estimated predicted outcome based on given data.
F statistics and p valuse for the general linear hypothesis
}
\description{
General Linear Hypothesis
}
\examples{
data(mtcars)
ma = GLH(x= list(mtcars$cyl,mtcars$hp,mtcars$drat,mtcars$wt), y = mtcars$mpg,intr = TRUE)
mb = GLH(x= list(mtcars$cyl,mtcars$hp,mtcars$drat,mtcars$wt), y = mtcars$mpg, intr = TRUE,contrast = matrix(c(0,0,1,-1,0),byrow = TRUE,nrow =1),rhs = 0.8)
}
